Add an option to customize request methods with body params parsing
This optional parameter allows to accept body params with request
methods other than the default POST, PUT, PATCH, and DELETE.

The main use case is accepting body params for GET requests.
Although it is not a very common practice, the HTTP spec allows
body data with GET requests. Elasticsearch for instance is a
product that utilises HTTP request bodies in GET.